GenAgent3D/requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.5.0
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0

# 3D and rendering
bpy>=3.0.0  # Blender Python API
pyrender>=0.1.45
trimesh>=3.9.0
open3d>=0.15.0

# Vision and ML
transformers>=4.30.0
clip>=1.0
torchvision>=0.15.0
pillow>=9.0.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
python-dotenv>=1.0.0
requests>=2.28.0

# Testing and development
pytest>=7.0.0
black>=22.0.0
isort>=5.10.0
flake8>=6.0.0 